changed simulator cpu to 64bit registers

+++ src/cpu/cpu.h
@@ -10,19 +10,19 @@ namespace hw
 {
 namespace constant 
 {
-static constexpr uint32_t REG_SIZE = 16;
-static constexpr uint32_t MEM_SIZE = 262144;
-static constexpr uint32_t DIS_WIDTH = 256;
-static constexpr uint32_t DIS_HEIGHT = 128;
-static constexpr uint32_t DIS_SCALE = 4;
+static constexpr uint64_t REG_SIZE = 16;
+static constexpr uint64_t MEM_SIZE = 262144;
+static constexpr uint64_t DIS_WIDTH = 256;
+static constexpr uint64_t DIS_HEIGHT = 128;
+static constexpr uint64_t DIS_SCALE = 4;
 }; // namespace constants
 
 struct Binary
 {
   std::vector<std::shared_ptr<Instruction>> instructions;
   std::unordered_map<std::string, size_t> label2pc;
-  uint32_t m_PC = 0;
-  uint32_t m_nextPC = 1;
+  uint64_t m_PC = 0;
+  uint64_t m_nextPC = 1;
 };
 
 struct CPU
@@ -37,10 +37,10 @@ struct CPU
   }
 
   Binary m_bin;
-  uint32_t m_regFile[constant::REG_SIZE] = {};
-  uint32_t m_mem[constant::MEM_SIZE] = {};
-  uint32_t m_run = 1;
-  uint32_t m_mem_idx = 0;
+  uint64_t m_regFile[constant::REG_SIZE] = {};
+  uint64_t m_mem[constant::MEM_SIZE] = {};
+  uint64_t m_run = 1;
+  uint64_t m_mem_idx = 0;
 };
 
 }; // namespace hw
